Moved default undo to BitmapOperationTool
